# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clientes-API",
        "description": "Em caso de <b>problemas/bugs</b> com alguma rota na API, por favor abra uma issue detalhando o erro e, se possível, forneça os passos necessários para reproduzi-lo.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "List clients",
                "description": "Returns a list of clients, filtered by name, email or phone, paginated by 10 items per page.",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "parameters": [
                    {
                        "name": "take",
                        "in": "query",
                        "description": "Selects a specified number of elements",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter clients by name (case insensitive).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter clients by email (case insensitive).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Filter clients by phone (case insensitive).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Carlos Henrique"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@gmail.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "99999999"
                                        },
                                        "address": {
                                            "properties": {
                                                "cep": {
                                                    "type": "string",
                                                    "example": "55730000"
                                                },
                                                "number": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "neighborhood": {
                                                    "type": "string",
                                                    "example": "Derby"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Bom Jardim"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "description": "Create a new client",
                "operationId": "f1e86e5b79fedfa6593067977292b13b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "address": {
                                                "properties": {
                                                    "cep": {
                                                        "type": "string"
                                                    },
                                                    "number": {
                                                        "type": "integer"
                                                    },
                                                    "neighborhood": {
                                                        "type": "string"
                                                    },
                                                    "city": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Carlos Henrique",
                                    "email": "email@gmail.com",
                                    "phone": "99999999",
                                    "user_id": 1,
                                    "address": {
                                        "cep": "55730000",
                                        "number": 5,
                                        "neighborhood": "Derby",
                                        "city": "Bom Jardim"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Carlos Henrique"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@gmail.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "99999999"
                                        },
                                        "address": {
                                            "properties": {
                                                "cep": {
                                                    "type": "string",
                                                    "example": "55730000"
                                                },
                                                "number": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "neighborhood": {
                                                    "type": "string",
                                                    "example": "Derby"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Bom Jardim"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}